split path_to_matrices to files
